% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_name_functions.R
\name{apply_name_fixes}
\alias{apply_name_fixes}
\title{Apply suggested name fixes to either phyloseq object or metadata}
\usage{
apply_name_fixes(
  fixes,
  physeq = NULL,
  metadata = NULL,
  target = c("physeq", "metadata")
)
}
\arguments{
\item{fixes}{A data frame of fixes, typically from suggest_name_fixes()}

\item{physeq}{A phyloseq object to rename samples in (if target="physeq")}

\item{metadata}{A data frame to rename samples in (if target="metadata")}

\item{target}{Whether to rename samples in "physeq" or "metadata"}
}
\value{
Updated phyloseq object or metadata data frame with corrected names
}
\description{
Apply suggested name fixes to either phyloseq object or metadata
}
\examples{
\dontrun{
data(GlobalPatterns, package = "phyloseq")
meta <- data.frame(sample_data(GlobalPatterns))
# Modify some names to create mismatches
rownames(meta)[1:3] <- paste0(rownames(meta)[1:3], "_modified")
suggestions <- suggest_name_fixes(GlobalPatterns, meta)
# Select which suggestions to apply
fixes_to_apply <- suggestions[1:2, ]
# Apply to metadata
fixed_meta <- apply_name_fixes(fixes_to_apply, metadata = meta, target = "metadata")
}
}
