% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_functions.R
\name{get_template_path}
\alias{get_template_path}
\title{Get the path to the phyloseq-meta-check template}
\usage{
get_template_path()
}
\value{
Path to the template
}
\description{
This function returns the path to the R Markdown template for phyloseq-meta-check
}
\keyword{internal}
