% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_phyloseq}
\alias{is_phyloseq}
\title{Check if an object is a phyloseq object}
\usage{
is_phyloseq(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Boolean indicating whether x is a phyloseq object
}
\description{
Check if an object is a phyloseq object
}
\keyword{internal}
