% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_metadata}
\alias{is_valid_metadata}
\title{Check if a data frame can be used as metadata}
\usage{
is_valid_metadata(x)
}
\arguments{
\item{x}{Data frame to check}
}
\value{
Boolean indicating whether x can be used as metadata
}
\description{
Checks if the data frame has row names that can be used as sample identifiers
}
\keyword{internal}
