% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_functions.R
\name{use_template}
\alias{use_template}
\title{Use phyloseq-meta-check template to create a new R Markdown document}
\usage{
use_template(file)
}
\arguments{
\item{file}{Path where the new file will be created}
}
\value{
Path to the created file (invisibly)
}
\description{
Use phyloseq-meta-check template to create a new R Markdown document
}
\examples{
\dontrun{
use_template("my_report.Rmd")
}
}
